Fix tg export --quilt --numbered w/ > 8 patches
Creating a quilt patch series with more than 8 patches with:
tg export --quilt --numbered patchdir

Causes the following error to be produced:
eval: 1: arithmetic expression: expecting EOF: "0008 + 1"

This is due to the leading 0 and the $(()) evaluates as octal rather
than decimal.

Add the leading zeros only to the filename, not the number itself.

Signed-off-by: Jon Ringle <jon@ringle.org>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>