Implemented "tg summary --sort".
This uses tsort to sort the branches based on their dependency graph.
Note: only a single sort order that fulfills the dependency graph is
printed.  There may be many other possible orderings.

A graphical text view would be much more useful, but this is till
a lot better than nothing.

Signed-off-by: Per Cederqvist <ceder@lysator.liu.se>
Signed-off-by: Uwe Kleine-König <u.kleine-koenig@pengutronix.de>