add an option to allow tg-export --collapse/--linearize to used with an existing branch
Add a new option, `--force`, to `tg-export`, such that if
`--collapse`/`--linearize` is used on an existing branch, that branch is
reset to the resulting export.

This is done by passing `-B` to `git-checkout`. The `-B` option has been
available in git since v1.7.3 (2010 Sept).  While its lowercase
companion `-b` creates a new branch and checks out to it, `-B` resets
the named branch if it already exists (`b` does not allow the named
branch to exist).

Signed-off-by: Tay Ray Chuan <rctay89@gmail.com>