Build improvments
Support an optional config.mak file the same way the
Git build does by including it with -include and adding
it to .gitignore.  This way changes to the build can
be made persistent without needing to remember the exact
options that have been set and without making the output
of git describe "dirty".

Track changes to prefix and related variables and
automatically rebuild when necessary just like Git does.