Fix precheck test
Adjust the precheck version test so that it is
insensitive to case, translation and suffixes that
may be present in the output of the `git version`
command.

Some versions of Git are known to output a version
that looks like 'git version n.n.n.n (foo foo)'.

Without this fix those versions of Git do not pass
the check even though they should if the n.n.n.n
part is sufficiently recent.